Fixed and added support for QwtLogScaleEngine (<=> QwtLog10ScaleEngine in Qwt5)

+++ qwt/qwt_point_mapper.py
@@ -100,8 +100,8 @@ def qwtToPointsFiltered(boundingRect, xMap, yMap, series, from_, to,
     points = []
     for i in range(from_, to+1):
         sample = series.sample(i)
-        x = round(xMap.transform(sample.x()))
-        y = round(yMap.transform(sample.y()))
+        x = int(round(xMap.transform(sample.x())))
+        y = int(round(yMap.transform(sample.y())))
         points.append(Point(x, y))
     return Polygon(list(set(points)))
